Reset connections
